requirements.txt
Django==4.2.7
django-ninja==1.5.0
django-ninja-jwt==5.4.2
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
structlog==23.2.0

